+++ package.json
@@ -112,6 +112,7 @@
     "react-intersection-observer": "^9.4.0",
     "react-lazy-load": "^4.0.1",
     "react-lowlight": "^3.0.0",
+    "react-progressive-graceful-image": "^0.7.0",
     "react-sticky-box": "^1.0.2",
     "react-sticky-el": "^2.0.9",
     "react-syntax-highlighter": "^15.5.0",
